불필요한 package depth 제거
