+++ src/main/java/com/daemawiki/external/api/RSocketConfiguration.java
@@ -1,0 +1,19 @@
+package com.daemawiki.external.api;
+
+import org.springframework.boot.rsocket.messaging.RSocketStrategiesCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.codec.json.Jackson2JsonDecoder;
+import org.springframework.http.codec.json.Jackson2JsonEncoder;
+
+@Configuration
+class RSocketConfiguration {
+
+    @Bean
+    RSocketStrategiesCustomizer rSocketStrategiesCustomizer() {
+        return strategies -> strategies
+            .encoder(new Jackson2JsonEncoder())
+            .decoder(new Jackson2JsonDecoder());
+    }
+
+}
