+++ src/main/java/com/daemawiki/external/api/annotation/DocumentsRestApi.java
@@ -1,0 +1,14 @@
+package com.daemawiki.external.api.annotation;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.lang.annotation.*;
+
+@Documented
+@RestController
+@RequestMapping("/api/documents")
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE})
+public @interface DocumentsRestApi {
+}
