+++ src/main/java/com/daemawiki/external/api/rest/auth/dto/LoginForm.java
@@ -1,0 +1,10 @@
+package com.daemawiki.external.api.rest.auth.dto;
+
+import com.daemawiki.internal.user.primitive.Password;
+import com.daemawiki.internal.user.primitive.personal.Email;
+
+public record LoginForm(
+        Email email,
+        Password password
+) {
+}
