+++ src/main/java/com/daemawiki/external/api/rest/document/DocumentDTOMapper.java
@@ -1,0 +1,59 @@
+package com.daemawiki.external.api.rest.document;
+
+import com.daemawiki.external.api.rest.document.dto.DocumentFullResponse;
+import com.daemawiki.external.api.rest.document.dto.DocumentHalfResponse;
+import com.daemawiki.external.api.rest.document.dto.DocumentCreateForm;
+import com.daemawiki.internal.document.dto.DocumentInternalDTO;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+
+import java.util.List;
+
+import static org.mapstruct.MappingConstants.ComponentModel.SPRING;
+import static org.mapstruct.ReportingPolicy.IGNORE;
+
+@Mapper(
+        componentModel = SPRING,
+        unmappedTargetPolicy = IGNORE,
+        unmappedSourcePolicy = IGNORE
+)
+public interface DocumentDTOMapper {
+
+    @Mapping(
+            target = "documentTitle",
+            source = "documentTitle",
+            ignore = false
+    )
+    @Mapping(
+            target = "categoryList",
+            source = "categoryList",
+            ignore = false
+    )
+    @Mapping(
+            target = "type",
+            source = "type",
+            ignore = false
+    )
+    DocumentInternalDTO toDocumentDTO(DocumentCreateForm source);
+
+    DocumentFullResponse toDocumentFullResponse(DocumentInternalDTO source);
+
+    @Mapping(
+            target = "mainTitle",
+            source = "documentTitle.mainTitle",
+            ignore = false
+    )
+    @Mapping(
+            target = "content",
+            source = "documentContent.textBody",
+            ignore = false
+    )
+    DocumentHalfResponse toDocumentHalfResponse(DocumentInternalDTO source);
+
+    default List<DocumentHalfResponse> toDocumentHalfResponseList(List<DocumentInternalDTO> source) {
+        return source.stream()
+                .map(this::toDocumentHalfResponse)
+                .toList();
+    }
+
+}
