+++ src/main/java/com/daemawiki/external/api/rest/document/dto/DocumentEditEditorSetForm.java
@@ -1,0 +1,10 @@
+package com.daemawiki.external.api.rest.document.dto;
+
+import com.daemawiki.internal.user.vo.Editor;
+
+import java.util.Set;
+
+public record DocumentEditEditorSetForm(
+        Set<Editor> editorList
+) {
+}
