+++ src/main/java/com/daemawiki/external/api/rest/document/dto/DocumentEditTitleForm.java
@@ -1,0 +1,8 @@
+package com.daemawiki.external.api.rest.document.dto;
+
+import com.daemawiki.internal.document.vo.DocumentTitle;
+
+public record DocumentEditTitleForm(
+        DocumentTitle documentTitle
+) {
+}
