--- src/main/java/com/daemawiki/external/rsocket/RSocketConfiguration.java
@@ -1,19 +1,0 @@
-package com.daemawiki.external.rsocket;
-
-import org.springframework.boot.rsocket.messaging.RSocketStrategiesCustomizer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.codec.json.Jackson2JsonDecoder;
-import org.springframework.http.codec.json.Jackson2JsonEncoder;
-
-@Configuration
-class RSocketConfiguration {
-
-    @Bean
-    RSocketStrategiesCustomizer rSocketStrategiesCustomizer() {
-        return strategies -> strategies
-            .encoder(new Jackson2JsonEncoder())
-            .decoder(new Jackson2JsonDecoder());
-    }
-
-}
