--- src/main/java/com/daemawiki/external/security/paseto/TokenUtils.java
@@ -1,14 +1,0 @@
-package com.daemawiki.external.security.paseto;
-
-import org.springframework.security.core.Authentication;
-import reactor.core.publisher.Mono;
-
-public interface TokenUtils {
-
-    Mono<Boolean> validateToken(String token);
-
-    Mono<String> extractUsername(String token);
-
-    Mono<Authentication> getAuthentication(String token);
-
-}
