--- src/main/java/com/daemawiki/external/web/annotation/ui/DocumentsRestApi.java
@@ -1,14 +1,0 @@
-package com.daemawiki.external.web.annotation.ui;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.lang.annotation.*;
-
-@Documented
-@RestController
-@RequestMapping("/api/documents")
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
-public @interface DocumentsRestApi {
-}
