--- src/main/java/com/daemawiki/external/web/annotation/ui/MailRestApi.java
@@ -1,18 +1,0 @@
-package com.daemawiki.external.web.annotation.ui;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Documented
-@RestController
-@RequestMapping("/api/mail")
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE})
-public @interface MailRestApi {
-}
