--- src/main/java/com/daemawiki/external/web/rest/auth/dto/ReissueResponse.java
@@ -1,9 +1,0 @@
-package com.daemawiki.external.web.rest.auth.dto;
-
-public record ReissueResponse(
-        String token
-) {
-    public static ReissueResponse create(String token) {
-        return new ReissueResponse(token);
-    }
-}
