--- src/main/java/com/daemawiki/external/web/rest/document/dto/form/DocumentEditEditorSetForm.java
@@ -1,10 +1,0 @@
-package com.daemawiki.external.web.rest.document.dto.form;
-
-import com.daemawiki.internal.user.vo.Editor;
-
-import java.util.Set;
-
-public record DocumentEditEditorSetForm(
-        Set<Editor> editorList
-) {
-}
