--- src/main/java/com/daemawiki/external/web/rest/document/dto/form/DocumentEditTitleForm.java
@@ -1,8 +1,0 @@
-package com.daemawiki.external.web.rest.document.dto.form;
-
-import com.daemawiki.internal.document.vo.DocumentTitle;
-
-public record DocumentEditTitleForm(
-        DocumentTitle documentTitle
-) {
-}
