+++ src/main/java/com/daemawiki/internal/user/LoginUseCase.java
@@ -1,6 +1,6 @@
 package com.daemawiki.internal.user;
 
-import com.daemawiki.external.web.rest.auth.dto.LoginResponse;
+import com.daemawiki.external.api.rest.auth.dto.LoginResponse;
 import com.daemawiki.internal.user.dto.LoginDTO;
 import reactor.core.publisher.Mono;
